+++ components/Header.js
@@ -2,20 +2,18 @@ import styled from "styled-components";
 import Navigation from "./Navigation";
 import { useState } from "react";
 import Logo from "../public/logo.svg";
+import Menu from "./../public/menu.svg";
+import Close from "./../public/close.svg";
 import Moon from "../public/moon.svg";
 import Sun from "../public/sun.svg";
 import { lightTheme, darkTheme } from "../components/Theme";
+import { StyledWrapper } from "@/SharedStyledComponents";
 
 const StyledToggleTheme = styled.button`
   border-radius: 50%;
   border: 1px solid var(--main-dark);
   background-color: transparent;
-  font-size: 1rem;
-  margin: 1rem;
-  position: fixed;
-  top: 0.7rem;
-  right: 4rem;
-  z-index: 3;
+  z-index: 2;
   aspect-ratio: 1/1;
 `;
 
@@ -27,9 +25,24 @@ const StyledSun = styled(Sun)`
   fill: var(--main-dark);
 `;
 
+const StyledMenuButton = styled(Menu)`
+  width: 2.5rem;
+  fill: var(--main-dark);
+  border-style: none;
+`;
+const StyledCloseButton = styled(Close)`
+  width: 2.5rem;
+  fill: var(--contrast-text);
+  border-style: none;
+  z-index: 2;
+`;
+
 const StyledHeader = styled.header`
   width: 100%;
   height: 100px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
   background: var(--main-bright);
   position: fixed;
   top: 0;
@@ -37,15 +50,20 @@ const StyledHeader = styled.header`
   z-index: ${({ $isOpen }) => ($isOpen ? "2" : "1")};
 `;
 
+const StyledIconWrapper = styled(StyledWrapper)`
+  justify-content: center;
+  align-items: center;
+  gap: 0.5rem;
+
+  width: auto;
+  padding: 1rem;
+`;
+
 const StyledLogo = styled(Logo)`
   width: 9rem;
   height: 9rem;
-  fill: var(--main-dark);
-  color: var(--main-dark);
+  padding-left: 0.5rem;
   stroke: var(--main-dark);
-  position: absolute;
-  top: -1.5rem;
-  left: 1rem;
 `;
 
 export default function Header({ theme, toggleTheme, switchTheme }) {
@@ -56,18 +74,27 @@ export default function Header({ theme, toggleTheme, switchTheme }) {
   }
 
   return (
-    <StyledHeader $isOpen={isOpen}>
-      <StyledLogo />
-      {theme === lightTheme || theme === darkTheme ? (
-        <StyledToggleTheme type="button" onClick={toggleTheme}>
-          {theme === lightTheme ? <StyledMoon /> : <StyledSun />}
-        </StyledToggleTheme>
-      ) : null}
-      <Navigation
-        isOpen={isOpen}
-        handleToggleMenu={handleToggleMenu}
-        switchTheme={switchTheme}
-      />
-    </StyledHeader>
+    <>
+      <StyledHeader $isOpen={isOpen}>
+        <StyledLogo />
+        <StyledIconWrapper>
+          {theme === lightTheme || theme === darkTheme ? (
+            <StyledToggleTheme type="button" onClick={toggleTheme}>
+              {theme === lightTheme ? <StyledMoon /> : <StyledSun />}
+            </StyledToggleTheme>
+          ) : null}
+          {isOpen ? (
+            <StyledCloseButton type="button" onClick={handleToggleMenu} />
+          ) : (
+            <StyledMenuButton type="button" onClick={handleToggleMenu} />
+          )}
+          <Navigation
+            isOpen={isOpen}
+            handleToggleMenu={handleToggleMenu}
+            switchTheme={switchTheme}
+          />
+        </StyledIconWrapper>
+      </StyledHeader>
+    </>
   );
 }
