+++ components/Theme.js
@@ -10,9 +10,9 @@ export const lightTheme = {
   contrastBright: "#f1eaea",
   red: "#cc0100",
   green: "#00b400",
-  joy: "hsl(60, 80%, 70%)",
-  surprise: "hsl(105, 80%, 80%)",
-  fear: "hsl(165, 100%, 70%)",
+  joy: "hsl(60, 70%, 70%)",
+  surprise: "hsl(105, 70%, 80%)",
+  fear: "hsl(165, 70%, 70%)",
   sadness: "hsl(200, 80%, 70%)",
   contempt: "hsl(265, 80%, 80%)",
   disgust: "hsl(310, 80%, 70%)",
